fix: 중복된 코드 분리
